Add TEST_NAMEDTENSOR flag to namedtensor ci (#25948)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25948

Previously, test/test_namedtensor.py is skipped if pytorch was not
compiled with BUILD_NAMEDTENSOR. Now, we skip test/test_namedtensor.py
if pytorch was not compiled with BUILD_NAMEDTENSOR or if
TEST_NAMEDTENSOR is not set.

This is done in preparation for turning on BUILD_NAMEDTENSOR=1 permanently;
at that point we will use TEST_NAMEDTENSOR to differentiate between the
named tensor ci and the regular ci.

Test Plan:
- [namedtensor ci] (and check that the named tensor tests are actually
running).

Differential Revision: D17300132

Pulled By: zou3519

fbshipit-source-id: 928f71f4d50445680b6ae1aa54b8857bc92e4d08