Puts ROCm tests on default stream (#26394)
Summary:
This PR has been updated. Since ORIGINAL PR comment below.

ROCm CI builds have been hanging as we've been refactoring tests, even when these refactors seem entirely innocuous. This PR started by commenting out test_stft, for example, a Python test never run on ROCm, and that was sufficient to reliably hang the ROCm build in CI.

Putting ROCm tests back on the default stream appears to remove this hang. So this PR now does that. This is likely to unblock development.

ORIGINAL: Some test changes appear to be causing ROCm builds to hang in CI. This PR is an attempt to diagnose the source of the hang.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26394

Test Plan: Change is to test themselves.

Differential Revision: D17456678

Pulled By: mruberry

fbshipit-source-id: 38d00d01c64b5055c1dfed01687ce3e1c9372887