Bump gloo (#27133)
Summary:
Includes a bugfix for the uv transport used on macOS.

See https://github.com/facebookincubator/gloo/pull/220 for details.

This is a re-submit of https://github.com/pytorch/pytorch/issues/27085 because it's not possible to merge PRs with submodule bumps that were submitted through ghstack at the moment.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27133

Differential Revision: D17685246

Pulled By: pietern

fbshipit-source-id: 260e963446036d63eb341a5b21ea717e9c7492a2