preserve original tensoriterator behavior when not explicitly promoting (#28231)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28231

Fix: https://github.com/pytorch/pytorch/issues/28010

A mixed-type index assignment that would have been an error in 1.2 was unintentionally made possible (with incorrect results) in 1.3. This PR restores the original behavior.

This is BC-breaking because:
```
        a = torch.ones(5, 2, dtype=torch.double)
        b = torch.zeros(5, dtype=torch.int)
        a[:, [1]] = b.unsqueeze(-1)
```
now raises an error (as in 1.2) whereas it did not in 1.3.

Test Plan: Imported from OSS

Differential Revision: D18049637

Pulled By: nairbv

fbshipit-source-id: 11a37dc98364ae70aac0e9dbc090d2a500aa7ccc